pyproject.toml
[project]
name = "rag-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.6",
    "groq>=1.0.0",
    "inngest>=0.5.15",
    "llama-index-core>=0.14.14",
    "llama-index-readers-file>=0.5.6",
    "openai>=2.20.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.3",
    "uvicorn>=0.40.0",
]
